Unify printing
Every struct or bit field change required multiple places to update
its display. Unify all nvme reflection through a single general purpose
object that can be printed in all the ways we want to display the various
records nvme-cli retrieves.

While the output does not entirely match the previous, this is
significantly easier to work with and we can improve the pretty-printing
as needed.

While doing this, it became apparent that some vendor specific plug-ins
had entirely copied the base functions and provided nothing of note that
requires vendor specific extentions. Delete those so we're not duplicating
code.

Signed-off-by: Keith Busch <kbusch@kernel.org>